+++ scripts/test.sh
@@ -16,4 +16,3 @@ go install ./...
 SQLFLOW_log_level=debug go test -v ./...  -covermode=count -coverprofile=coverage.out
 
 python -m unittest discover -v sql/python "*_test.py"
-python -m unittest discover -v sql/python "instance_db_test_.py"
