+++ scripts/test_hive.sh
@@ -25,4 +25,4 @@ go get -v -t ./...
 go install ./...
 SQLFLOW_log_level=debug go test -v ./...
 
-python -m unittest discover -v sql/python "instance_db_test_.py"
+python -m unittest discover -v sql/python "db_test_.py"
